+++ changelog.md
@@ -32,6 +32,7 @@
 * Tweak - Change WooCommerce Products Addon visual on single
 * Tweak - Add WooCommerce Bookings compatibility
 * Tweak - Add WooCommerce Product Reviews compatibility
+* Tweak - Add WooCommerce thank you page template
 
 ### 2.0.0
 * New totally awesome version! The visual was fully redesigned